+++ packages/nextjs-contentful-renderer/src/MdxRenderer/MarkdownContext/MarkdownContext.tsx
@@ -1,17 +1,5 @@
-import { LinkProps, TextProps } from '@cmpsr/components';
 import React, { PropsWithChildren, createContext, useContext } from 'react';
-
-interface MarkdownContextProps {
-  paragraph?: TextProps;
-  link?: LinkProps;
-  h1?: TextProps;
-  h2?: TextProps;
-  h3?: TextProps;
-  h4?: TextProps;
-  h5?: TextProps;
-  h6?: TextProps;
-  p?: TextProps;
-}
+import { MarkdownContextProps } from './types';
 
 const MarkdownContext = createContext<MarkdownContextProps>({});